+++ plugin-code/src/main/java/com/trivago/cluecumber/rendering/pages/renderering/AllStepsPageRenderer.java
@@ -50,16 +50,16 @@ public class AllStepsPageRenderer extends PageWithChartRenderer {
 
     private void addChartJsonToReportDetails(final AllStepsPageCollection allTagsPageCollection) {
 
-        List<Integer> passed = new ArrayList<>();
-        List<Integer> failed = new ArrayList<>();
-        List<Integer> skipped = new ArrayList<>();
+        List<Float> passed = new ArrayList<>();
+        List<Float> failed = new ArrayList<>();
+        List<Float> skipped = new ArrayList<>();
 
         int maximumNumberOfRuns = 0;
         for (Map.Entry<Step, ResultCount> entry : allTagsPageCollection.getStepResultCounts().entrySet()) {
             ResultCount value = entry.getValue();
-            passed.add(value.getPassed());
-            failed.add(value.getFailed());
-            skipped.add(value.getSkipped());
+            passed.add((float) value.getPassed());
+            failed.add((float) value.getFailed());
+            skipped.add((float) value.getSkipped());
             if (value.getTotal() > maximumNumberOfRuns) {
                 maximumNumberOfRuns = value.getTotal();
             }
